Support sandboxed execution on Linux
Add namespace isolation and subreaper-owned cleanup for default serve and
standalone sandbox commands. Share runner setup across platforms and stage
the native Linux runner and bubblewrap in wheels.

Resolve namespace PIDs before fixture cleanup, reject broadcast signal
IDs, and exercise Linux sandbox behavior through public transcript tests.

+++ src/cli.rs
@@ -79,9 +79,9 @@ pub enum Command {
     /// Restore the target signal mask and execute its command
     #[command(hide = true)]
     SandboxTarget {
-        /// Original macOS signal mask, encoded as an unsigned decimal integer
+        /// Original signal mask, encoded as an unsigned decimal integer
         #[arg(long, value_name = "MASK")]
-        signal_mask: u32,
+        signal_mask: u64,
 
         /// Command and arguments to run after host supervision is ready
         #[arg(
